+++ test/tasks/ods-build-gradle_test.go
@@ -25,6 +25,7 @@ func TestTaskODSBuildGradle(t *testing.T) {
 					ctxt.ODS = tasktesting.SetupGitRepo(t, ctxt.Namespace, wsDir)
 					ctxt.Params = map[string]string{
 						"sonar-quality-gate": "true",
+						"cache-build":        "false",
 					}
 				},
 				WantRunSuccess: true,
@@ -61,7 +62,7 @@ func TestTaskODSBuildGradle(t *testing.T) {
 					ctxt.ODS = tasktesting.SetupGitRepo(t, ctxt.Namespace, wsDir)
 					ctxt.Params = map[string]string{
 						"sonar-quality-gate": "true",
-						"cache-build":        "true",
+						// "cache-build":        "true", is expected to be default
 					}
 				},
 				WantRunSuccess: true,